requirements.txt
langchain-openai==0.0.3
langgraph==0.0.11
newsapi-python==0.2.7
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.5.2
aiohttp>=3.11.0,<3.12.0
pytest==7.4.3
pytest-asyncio==0.23.2
